+++ packages/graphql/tests/integration/directives/cypher.int.test.ts
@@ -24,7 +24,6 @@ import neo4j from "../neo4j";
 import { Neo4jGraphQL } from "../../../src/classes";
 import { createJwtRequest } from "../../utils/create-jwt-request";
 
-const testLabel = generate({ charset: "alphabetic" });
 describe("cypher", () => {
     let driver: Driver;
 
@@ -993,110 +992,6 @@ describe("cypher", () => {
                     expectedTownDestinationList
                 );
             });
-        });
-        // Reproduces https://github.com/neo4j/graphql/issues/630
-        test("should query nested connection", async () => {
-            const typeDefs = `
-                type Actor {
-                    id: ID!
-                    name: String!
-                    movies: [Movie!]! @cypher(statement: "MATCH (this)-[:ACTED_IN]->(m:Movie) RETURN m")
-                }
-
-                type Movie {
-                    id: ID!
-                    title: String!
-                    actors: [Actor!]! @relationship(type: "ACTED_IN", direction: IN)
-                }
-            `;
-
-            const { schema } = new Neo4jGraphQL({ typeDefs, driver });
-
-            const actors = [
-                {
-                    id: generate(),
-                    name: "Keanu Reeves",
-                },
-                {
-                    id: generate(),
-                    name: "Carrie-Ann Moss",
-                },
-            ];
-
-            const movie = {
-                id: generate(),
-                title: "The Matrix",
-            };
-
-            const session = driver.session();
-            try {
-                await session.run(
-                    `
-                    CREATE (movie:Movie:${testLabel}) SET movie = $movie
-                    CREATE (actor1:Actor:${testLabel}) SET actor1 = $actors[0]
-                    CREATE (actor2:Actor:${testLabel}) SET actor2 = $actors[1]
-                    MERGE (actor1)-[:ACTED_IN]->(movie)<-[:ACTED_IN]-(actor2)
-                `,
-                    { actors, movie }
-                );
-
-                const source = `
-                    query($actorId: ID!) {
-                        actors(where: { id: $actorId }) {
-                            id
-                            name
-                            movies {
-                                id
-                                title
-                                actorsConnection {
-                                    totalCount
-                                    edges {
-                                        node {
-                                            id
-                                            name
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                `;
-
-                const gqlResult = await graphql({
-                    schema,
-                    source,
-                    contextValue: { driver },
-                    variableValues: { actorId: actors[0].id },
-                });
-
-                expect(gqlResult.errors).toBeUndefined();
-
-                const gqlActor = gqlResult.data?.actors[0];
-
-                expect(gqlActor).toBeDefined();
-                expect(gqlActor).toEqual({
-                    ...actors[0],
-                    movies: [
-                        {
-                            ...movie,
-                            actorsConnection: {
-                                totalCount: 2,
-                                edges: expect.arrayContaining([
-                                    {
-                                        node: actors[0],
-                                    },
-                                    {
-                                        node: actors[1],
-                                    },
-                                ]),
-                            },
-                        },
-                    ],
-                });
-            } finally {
-                await session.run(`MATCH (node:${testLabel}) DETACH DELETE node`);
-                await session.close();
-            }
         });
     });
 });
